Endpoint confirmation modals

+++ src/frontend/app/shared/components/list/list-types/endpoint/endpoints-list-config.service.ts
@@ -19,6 +19,8 @@ import { EndpointModel, endpointStoreNames } from '../../../../../store/types/en
 import { EntityMonitorFactory } from '../../../../monitors/entity-monitor.factory.service';
 import { InternalEventMonitorFactory } from '../../../../monitors/internal-event-monitor.factory';
 import { PaginationMonitorFactory } from '../../../../monitors/pagination-monitor.factory';
+import { ConfirmationDialogConfig } from '../../../confirmation-dialog.config';
+import { ConfirmationDialogService } from '../../../confirmation-dialog.service';
 import { ITableColumn } from '../../list-table/table.types';
 import { IListAction, IListConfig, ListViewTypes } from '../../list.component.types';
 import { EndpointsDataSource } from './endpoints-data-source';
@@ -27,7 +29,6 @@ import { TableCellEndpointStatusComponent } from './table-cell-endpoint-status/t
 
 import { map, pairwise } from 'rxjs/operators';
 import { Observable } from 'rxjs';
-
 
 
 function getEndpointTypeString(endpoint: EndpointModel): string {
@@ -89,9 +90,17 @@ export const endpointColumns: ITableColumn<EndpointModel>[] = [
 export class EndpointsListConfigService implements IListConfig<EndpointModel> {
   private listActionDelete: IListAction<EndpointModel> = {
     action: (item) => {
-      this.store.dispatch(new UnregisterEndpoint(item.guid, item.cnsi_type));
-      this.handleDeleteAction(item, ([oldVal, newVal]) => {
-        this.store.dispatch(new ShowSnackBar(`Unregistered ${item.name}`));
+      const confirmation = new ConfirmationDialogConfig(
+        'Unregister Endpoint',
+        `Are you sure you want to unregister endpoint '${item.name}'?`,
+        'Unregister',
+        true
+      );
+      this.confirmDialog.open(confirmation, () => {
+        this.store.dispatch(new UnregisterEndpoint(item.guid, item.cnsi_type));
+        this.handleDeleteAction(item, ([oldVal, newVal]) => {
+          this.store.dispatch(new ShowSnackBar(`Unregistered ${item.name}`));
+        });
       });
     },
     label: 'Unregister',
@@ -101,10 +110,18 @@ export class EndpointsListConfigService implements IListConfig<EndpointModel> {
 
   private listActionDisconnect: IListAction<EndpointModel> = {
     action: (item) => {
-      this.store.dispatch(new DisconnectEndpoint(item.guid, item.cnsi_type));
-      this.handleUpdateAction(item, EndpointsEffect.disconnectingKey, ([oldVal, newVal]) => {
-        this.store.dispatch(new ShowSnackBar(`Disconnected ${item.name}`));
-        this.store.dispatch(new GetSystemInfo());
+      const confirmation = new ConfirmationDialogConfig(
+        'Disconnect Endpoint',
+        `Are you sure you want to disconnect endpoint '${item.name}'?`,
+        'Disconnect',
+        false
+      );
+      this.confirmDialog.open(confirmation, () => {
+        this.store.dispatch(new DisconnectEndpoint(item.guid, item.cnsi_type));
+        this.handleUpdateAction(item, EndpointsEffect.disconnectingKey, ([oldVal, newVal]) => {
+          this.store.dispatch(new ShowSnackBar(`Disconnected ${item.name}`));
+          this.store.dispatch(new GetSystemInfo());
+        });
       });
     },
     label: 'Disconnect',
@@ -180,7 +197,8 @@ export class EndpointsListConfigService implements IListConfig<EndpointModel> {
     private paginationMonitorFactory: PaginationMonitorFactory,
     private entityMonitorFactory: EntityMonitorFactory,
     private internalEventMonitorFactory: InternalEventMonitorFactory,
-    private currentUserPermissionsService: CurrentUserPermissionsService
+    private currentUserPermissionsService: CurrentUserPermissionsService,
+    private confirmDialog: ConfirmationDialogService
   ) {
     this.dataSource = new EndpointsDataSource(
       this.store,
